Benchmark comparison taken from https://github.com/whyrusleeping/cbor-gen/pull/67#issuecomment-1076547989 

```
name                       old time/op       new time/op       delta
SerializeNode-8                 9.88µs ±13%      10.92µs ± 5%  +10.54%  (p=0.009 n=10+10)
GetNode-8                       17.8µs ± 2%       14.4µs ± 5%  -19.13%  (p=0.000 n=10+10)
Find/n=1k/bitwidth=5-8          24.0µs ± 2%       23.0µs ± 2%   -3.98%  (p=0.000 n=10+10)
Find/n=5k/bitwidth=5-8          28.3µs ± 3%       27.7µs ± 5%     ~     (p=0.065 n=9+10)
Find/n=10k/bitwidth=5-8         23.1µs ± 3%       21.7µs ± 2%   -6.07%  (p=0.000 n=10+10)
Find/n=50k/bitwidth=5-8         38.4µs ± 7%       36.1µs ± 6%   -5.83%  (p=0.015 n=10+10)
Find/n=100k/bitwidth=5-8        41.0µs ± 8%       41.2µs ± 6%     ~     (p=1.000 n=10+10)
Find/n=500k/bitwidth=5-8        30.2µs ± 3%       29.3µs ± 4%   -3.29%  (p=0.003 n=10+10)
Find/n=1000k/bitwidth=5-8       36.7µs ± 2%       35.3µs ± 1%   -3.91%  (p=0.000 n=9+8)

name                       old alloc/op      new alloc/op      delta
SerializeNode-8                 14.9kB ± 3%       16.3kB ± 3%   +9.57%  (p=0.000 n=10+10)
GetNode-8                       20.7kB ± 0%       18.7kB ± 0%   -9.61%  (p=0.000 n=10+10)
Find/n=1k/bitwidth=5-8          35.0kB ± 2%       34.7kB ± 1%     ~     (p=0.156 n=10+9)
Find/n=5k/bitwidth=5-8          35.6kB ± 4%       35.4kB ± 6%     ~     (p=0.631 n=10+10)
Find/n=10k/bitwidth=5-8         23.3kB ± 2%       22.5kB ± 2%   -3.20%  (p=0.000 n=10+10)
Find/n=50k/bitwidth=5-8         51.2kB ± 0%       50.3kB ± 1%   -1.73%  (p=0.000 n=10+10)
Find/n=100k/bitwidth=5-8        55.0kB ± 1%       54.1kB ± 1%   -1.69%  (p=0.000 n=10+10)
Find/n=500k/bitwidth=5-8        33.5kB ± 0%       32.7kB ± 0%   -2.33%  (p=0.000 n=10+10)
Find/n=1000k/bitwidth=5-8       46.3kB ± 0%       45.3kB ± 0%   -2.06%  (p=0.000 n=10+10)

name                       old allocs/op     new allocs/op     delta
SerializeNode-8                    111 ± 2%          107 ± 5%   -3.86%  (p=0.001 n=10+10)
GetNode-8                          519 ± 0%          264 ± 0%  -49.13%  (p=0.000 n=10+10)
Find/n=1k/bitwidth=5-8             478 ± 2%          398 ± 1%  -16.63%  (p=0.000 n=10+10)
Find/n=10k/bitwidth=5-8            413 ± 1%          330 ± 1%  -20.25%  (p=0.000 n=10+10)
Find/n=50k/bitwidth=5-8            728 ± 0%          600 ± 0%  -17.50%  (p=0.000 n=10+9)
Find/n=100k/bitwidth=5-8           784 ± 0%          647 ± 0%  -17.45%  (p=0.000 n=10+10)
Find/n=500k/bitwidth=5-8           600 ± 0%          479 ± 0%  -20.17%  (p=0.000 n=10+10)
Find/n=1000k/bitwidth=5-8          743 ± 0%          602 ± 0%  -19.04%  (p=0.000 n=9+10)
```